==> notebooks/1_data_collection_Encode.ipynb <==
## Aim: Collect and combine relevant mouse data from the ENCODE data portal 

Link: https://www.encodeproject.org. 

From the web portal, we used a number of filters to retrieve the transcriptome data of healthy mouse blood cells. 

The filters we used are listed are as follows: 

- Assay type: Transcription (+)
- Organism: Mus musculus (+)
- Available file types: tsv (+)
- Biosample classification: cell line (-)
- Cell: leukocyte (+), hematopoietic cell (+), T cell (+), myeloid cell (+), monocyte (+), stem cell (+), progenitor cell (+)
- Audit category: low replicate concordance (-)

where +/- indicate positive or negative filters

### Samples

After the filtering, 90 samples from 38 experiments were selected. 

These selected samples came from two projects:

- ENCODE
- GGR

We downloaded the transcriptome data along with the associated meatadata. 

Note that the transcriptome data is downloaded as 90 individual data files, each corresponding to a single sample and named with an unique file accession. 

We will need to integrate the separate data files into a combined dataset. 

### Processing steps 

- Metadata handling
- Merge data of ENCODE project 
- Merge data of GGR project
- Combine Encode and GGR data 

import pandas as pd
import numpy as np
import atlas

# Load the metadata
meta_exp       = pd.read_csv('../data/raw/encode/experiment_metadata.tsv', sep='\t', index_col=0, skiprows=[0])
meta_exp

Note that, each entry of the metadata corresponds to an experiment, which give rise to a variety kinds of data besides the gene quantification file of our interest (e.g. read files in fastq). The index of the metadata is the experiment accession, which does not directly link to the file accessions of our downloaded data. Thus, we will need to transform the original metadata so that the index of the metadata are file accessions and those irrelavant files are excluded. 

In terminal: 

`ls -d ENCFF* | cut -f 1 -d '.' > file_accessions.txt` to record all the selected file accessions into a file called 'file_accessions.txt'. 

__Convert the index of the metadata from experiment ID to file accession__

meta_exp.Files = meta_exp.Files.str.split("[\\s*,]+") # convert the File field from comma separated strings to lists of strings

# load the accessions of the selected gene quantification files.
with open('../data/raw/encode/hiseq2000_files.tsv', 'r') as f:
    hiseq2000_files = f.read().splitlines()[2:]
    
# add platform variable to the metadata 

meta_exp['Platform'] = 'Illumina NextSeq'
meta_exp.loc[meta_exp.index.isin(hiseq2000_files), ['Platform']] = 'Illumina HiSeq'

# load the accessions of the selected gene quantification files.
with open('../data/raw/encode/file_accessions.txt', 'r') as f:
    selected_files = f.read().splitlines()

lst_col = 'Files'

meta_file = pd.DataFrame({
      col:np.repeat(meta_exp[col].values, meta_exp[lst_col].str.len())
      for col in meta_exp.columns.drop(lst_col)}
    ).assign(**{lst_col:np.concatenate(meta_exp[lst_col].values)})[meta_exp.columns]
meta_file.Files = meta_file.Files.replace(regex={r'^/files/': '', r'/$':''}) # remove prefix and suffix of the Files field
meta_file = meta_file.loc[meta_file.Files.isin(selected_files)].set_index('Files')

meta_file

# Example ENCODE sample
pd.read_csv('../data/raw/Encode/ENCFF040EGE.tsv', sep='\t', index_col=0).head()

'../data/raw/Encode/ENCFF547CXK.tsv'

# Example GGR sample
pd.read_csv('../data/raw/Encode/ENCFF022JXA.tsv', sep='\t', index_col=0, names=['ensembl','count']).head()

We can observe that gene quantification files generated from GGR project and ENCODE project have differnt formats. Thus we need to deal with the data generated from the two projects separately. 

### Merge expression data derived from ENCODE project  (41 samples)

# Retrieve accessions of files generated from the ENCODE project
encode_files = meta_file[meta_file.Project=='ENCODE'].index

def valid_rows(df):
    df = df[df.index.str.startswith("ENSMUSG")]
    df.index = df.index.str.replace(r'\.\d+', '') # remove the suffix of gene ID which represents the gene version
    return df 

def encode_file_processor(filenames):
    merged = pd.DataFrame()
    for name in filenames:
        df = pd.read_csv('../data/raw/Encode/'+name+'.tsv', sep='\t', index_col=0)
        df = valid_rows(df)[['TPM']]
        df.columns = [name]
        if merged.empty is True:
            merged = df
        else:
            merged = pd.merge(merged, df, how='inner', left_index=True, right_index=True)
    return(merged)  

# Merge expression data 
expression_encode = encode_file_processor(encode_files)
expression_encode.head()  

### Merge expression data derived from GGR project (49 samples)

GGR_files = meta_file[meta_file.Project=='GGR'].index

def GGR_file_processor(filenames):
    merged = pd.DataFrame()
    for name in filenames:
        df = pd.read_csv('../data/raw/encode/'+name+'.tsv', sep='\t', names = ["gene_id", "count"], index_col=0)
        df = valid_rows(df)
        df.columns = [name]
        if merged.empty is True:
            merged = df
        else:
            merged = pd.merge(merged, df, how='inner', left_index=True, right_index=True)
    return(merged)  

# Merge the count data of GGR samples 
GGR_files = meta_file[~meta_file.index.isin(encode_files)].index
expression_GGR = GGR_file_processor(GGR_files)
expression_GGR.head()

### Combine the two project together into a merged dataset

merged_dataset = pd.merge(expression_encode, expression_GGR, how='inner', left_index=True, right_index=True)
merged_dataset.head()

meta_file.head()

# Save the combined dataset
merged_dataset.to_csv('../data/interim/mouse_integrate/expression_encode.tsv', sep='\t')
meta_file.to_csv('../data/interim/mouse_integrate/samples_encode.tsv', sep='\t')
