==> notebooks/1_data_collection_Haemosphere.ipynb <==
## Aim: Collect and combine relevant mouse expression data from the Haemosphere data portal 

Link: https://haemosphere.org/datasets/show

Haemosphere is an online data portal that contains many gene expression datasets that focused on haematopoeisis in mice and humans. 

### Datasets 

We selected 6 datasets:

|__No.__|__dataset name__|__platform__|__platform details__|
|-------|----------------|------------|--------------------|
|1|Haemopedia-Mouse-RNAseq|RNAseq|HiSeq|
|2|Immgen-ULI-RNAseq|RNAseq|Illumina NextSeq 500|
|3|Haemopedia|microarray|Illumina WG6|
|4|Gene-Expression-Commons|microarray|Affymetrix Mouse430_2|
|5|Immgen|microarray|Affymetrix Mouse Gene 1.0 ST Array|
|6|Goodell|microarray|Affymetrix Mouse430_2|

import pandas as pd
import atlas
import handler
import GEOparse

# Load data

# Haemopedia RNAseq dataset 
expression1 = pd.read_csv('../data/raw/Haemosphere/Haemopedia-Mouse-RNASeq/Haemopedia-Mouse-RNASeq_tpm.txt', sep='\t', index_col=0)
samples1    = pd.read_csv('../data/raw/Haemosphere/Haemopedia-Mouse-RNASeq/Haemopedia-Mouse-RNASeq_samples.txt', sep='\t', index_col=0)

# Immgen ULI dataset 
expression2 = pd.read_csv('../data/raw/Haemosphere/Immgen-RNAseq/Immgen-ULI-RNASeq_tpm.txt', sep='\t', index_col=0)
samples2    = pd.read_csv('../data/raw/Haemosphere/Immgen-RNAseq/Immgen-ULI-RNASeq_samples.txt', sep='\t', index_col=0)

# Haemopedia microarray dataset 
expression3 = pd.read_csv('../data/raw/Haemosphere/Haemopedia-microarray/Haemopedia_expression.txt', sep='\t', index_col=0)
samples3    = pd.read_csv('../data/raw/Haemosphere/Haemopedia-microarray/Haemopedia_samples.txt', sep='\t', index_col=0)
probes3     = pd.read_csv('../data/raw/Haemosphere/Haemopedia-microarray/Haemopedia_probes.txt', sep='\t', index_col=0)

# Gene expression commons dataset
expression4 = pd.read_csv('../data/raw/Haemosphere/Gene-Expression-Commons/Gene-Expression-Commons_expression.txt', sep='\t', index_col=0)
samples4    = pd.read_csv('../data/raw/Haemosphere/Gene-Expression-Commons/Gene-Expression-Commons_samples.txt', sep='\t', index_col=0)
probes4     = pd.read_csv('../data/raw/Haemosphere/Gene-Expression-Commons/Gene-Expression-Commons_probes.txt', sep='\t', index_col=0)

# Immgen microarray dataset
expression5 = pd.read_csv('../data/raw/Haemosphere/Immgen-microarray/Immgen_expression.txt', sep='\t', index_col=0)
samples5    = pd.read_csv('../data/raw/Haemosphere/Immgen-microarray/Immgen_samples.txt', sep='\t', index_col=0)
probes5     = pd.read_csv('../data/raw/Haemosphere/Immgen-microarray/Immgen_probes.txt', sep='\t', index_col=0)

# Goodell dataset
expression6 = pd.read_csv('../data/raw/Haemosphere/Goodell-microarray/Goodell_expression.txt', sep='\t', index_col=0)
samples6    = pd.read_csv('../data/raw/Haemosphere/Goodell-microarray/Goodell_samples.txt', sep='\t', index_col=0)
probes6     = pd.read_csv('../data/raw/Haemosphere/Goodell-microarray/Goodell_probes.txt', sep='\t', index_col=0)

# For sample metadata contains normalised sample ids such as GEO_accessions and 
# SRA_accessions, use them as sampleId.
expression1 = pd.merge(expression1.T, samples1['sampleCode'], left_index=True, right_index=True).set_index('sampleCode').T
samples1 = samples1.set_index('sampleCode')

expression2 = pd.merge(expression2.T, samples2['sample_geo_accession'], left_index=True, right_index=True).set_index('sample_geo_accession').T
samples2 = samples2.set_index('sample_geo_accession')

### Merge selected expression datasets

def replace_probes(df, probe_mapping):
    '''
    Input: A microarray expression dataframe and and a probe mapping table.
    Output: The expression dataframe with index changed from probe ids to ensembl gene ids according to the supplied mapping table.
    '''
    probe_mapping.columns = ['ensembl']
    return pd.merge(df, probe_mapping, how='inner', left_index=True, right_index=True).set_index('ensembl')

# For microarray datasets, replace index of expression table from probe id to gene id
expression3 = replace_probes(expression3, probes3)
expression4 = replace_probes(expression4, probes4)
expression5 = replace_probes(expression5, probes5)
expression6 = replace_probes(expression6, probes6)

# merge samples of the 6 selected datasets
dfs = [expression1, expression2, expression3, expression4, expression5, expression6]
common_genes = handler.find_common_genes(dfs)
merged_ds = handler.merge_columns(dfs, common_genes)
expression_Haemosphere = atlas.rankTransform(merged_ds)

print(expression_Haemosphere.shape)
expression_Haemosphere.head()

### Standardise the metadata of each dataset

Since the sample metadata of different datasets may have recorded differnet variables, or using different names for similar attributes, to merge them, we identified a minimal list of necessary attributes to describe samples, and unified the naming of these attributes. The essential list of attributes includes:


- cell type
- cell lineage 
- description 
- markers 
- dataset_name  
- platform  


# metadata of Haemopedia RNAseq dataset
samples1.head() 

# metadata of Haemopedia RNAseq dataset - after standardisation
samples1_std = samples1[['celltype', 'cell_lineage', 'cell Type Description', 'immunophenotype']]
samples1_std.columns = ['celltype', 'cell_lineage', 'description', 'markers']
samples1_std['platform'] = 'RNAdeq (Illumina HiSeq)'
samples1_std['dataset_name'] = 'Haemopedia-Mouse-RNAseq'
samples1_std.head()

# metadata of Immgen ULI RNAseq dataset
samples2.head()

# metadata of Immgen ULI RNAseq dataset - after standardisation 
samples2_std = samples2[['celltype', 'cell_lineage', 'description', 'sort markers']]
samples2_std.columns = ['celltype', 'cell_lineage', 'description', 'markers']
samples2_std['platform'] = 'RNAseq (Illumina NextSeq)'
samples2_std['dataset_name'] = 'Immgen-ULI-RNAseq'
samples2_std.head()

# metadata of Haemopedia microarray dataset 
samples3.head()

# metadata of Haemopedia microarray dataset - after standardisation  
samples3_std = samples3[['celltype', 'cell_lineage', 'description', 'surface_markers']]
samples3_std.columns = ['celltype', 'cell_lineage', 'description', 'markers']
samples3_std['platform'] = 'Microarray (Illumina WG6)'
samples3_std['dataset_name'] = 'Haemopedia-microarray'
samples3_std.head()

# metadata of Gene-Expression-Commons dataset 
samples4.head()

# metadata of Gene-Expression-Commons dataset - after standardisation 
samples4_std = samples4[['celltype', 'cell_lineage', 'description', 'surface_markers']]
samples4_std.columns = ['celltype', 'cell_lineage', 'description', 'markers']
samples4_std['platform'] = 'Microarray (Affymetrix Mouse430_2)'
samples4_std['dataset_name'] = 'Gene-Expression-Commons'
samples4_std.head()

# metadata of Immgen microarray dataset
samples5.head()

# metadata of Gene-Expression-Commons dataset - after standardisation 
samples5_std = samples5[['celltype', 'cell_lineage']]
samples5_std['description'] = samples5.celltype
samples5_std['platform'] = 'Microarray (Affymetrix Mouse Gene 1.0 ST Array)'
samples5_std['dataset_name'] = 'Immgen-microarray'
samples5_std.head()

# metadata of Goodell dataset
samples6.head()

# metadata of Goodell dataset - after standardisation 
samples6_std = samples6[['celltype', 'cell_lineage', 'description', 'surface_markers']]
samples6_std.columns = ['celltype', 'cell_lineage', 'description', 'markers']
samples6_std['platform'] = 'Microarray (Affymetrix Mouse430_2)'
samples6_std['dataset_name'] = 'Goodell'
samples6_std.head()

### Merge standardised sample metadata 

samples_Haemosphere = pd.concat([samples1_std, samples2_std, samples3_std, samples4_std, samples5_std, samples6_std])
print(samples_Haemosphere.shape)
samples_Haemosphere.head()

### Sample selection
Remove non-blood samples

# inspect cell_lineage information of the combined sample metadata
set(samples_Haemosphere.cell_lineage)

'Non-haematopoietic' and 'Stromal cells' are non-blood samples and should be removed.

# Remove samples with cell_lineage annotated with 'Non-haematopoietic' and 'Stromal cells'.
samples_Haemosphere = samples_Haemosphere[~samples_Haemosphere.cell_lineage.isin(['Non-haematopoietic', 'Stromal Cells'])]
expression_Haemosphere = expression_Haemosphere.loc[:, expression_Haemosphere.columns.isin(samples_Haemosphere.index)]
expression_Haemosphere.shape

samples_Haemosphere.to_csv('../data/interim/mouse_integrate/samples_Haemosphere.tsv', sep='\t')
expression_Haemosphere.to_csv('../data/interim/mouse_integrate/expression_Haemosphere.tsv', sep='\t')
